클라이언트 ForgeUpgradeData, TownUpgradeData 부분 수정 - Forgeg확장

+++ Client/Assets/@Scripts/Data/Data.Contents.cs
@@ -150,8 +150,8 @@ namespace Data
         public int TemplateId;
         public string UpgradeName;
         public long Price;
-        public int CurrentValue;
-        public int NextValue;
+        public long CurrentValue;
+        public long NextValue;
         public int OriginalTemplateId;
         public int NextTempalteId;
